Replace size_t in the API with fixed-length types
For an API to be stable and predictable, its types
should be the same regardless of which platform
a node was built on.

size_t is different on x86 / x64.

Fixes #169

+++ source/agora/network/NetworkClient.d
@@ -219,7 +219,7 @@ class NetworkClient
 
     ***************************************************************************/
 
-    public const(Block)[] getBlocksFrom (ulong block_height, size_t max_blocks)
+    public const(Block)[] getBlocksFrom (ulong block_height, uint max_blocks)
     {
         return this.attemptRequest(
             this.api.getBlocksFrom(block_height, max_blocks), this.exception);
